Add endpoints in question run with token verification

+++ soa/question-run/src/question-run/question-run.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { NestApplicationContextOptions } from '@nestjs/common/interfaces/nest-application-context-options.interface';
 import { CreateAnswerDto } from 'src/answer/dto/create-answer.dto';
 import axios from "axios";
+import { verify } from 'src/general/gen_functions'
 import { BadRequestException } from '@nestjs/common';
 import { UpdateAnswerDto } from 'src/answer/dto/update-answer.dto';
 import { CreateUserAnswerVoteDto } from 'src/user-answer-vote/dto/create-user-answer-vote.dto';
@@ -10,31 +11,77 @@ axios.defaults.baseURL = 'http://localhost:3030';
 
 @Injectable()
 export class QuestionRunService {
-    createAnswer(body: CreateAnswerDto){
-        const requestUrl = 'answer';
-        return axios.post(requestUrl,body).then((response) => {return response.data;})
-        .catch(() => {
-            throw new  BadRequestException("Could not fetch data from the Data Layer.");
-        });
+    async createAnswer(headers:any, body: CreateAnswerDto){
+        let id : number = await verify(headers);
+
+        if(body.user.id == id) {
+            const requestUrl = 'answer';
+            return axios.post(requestUrl,body).then((response) => {return response.data;})
+            .catch(() => {
+                throw new  BadRequestException("Could not fetch data from the Data Layer.");
+            });
+        }
+        else {
+            throw new UnauthorizedException("Unauthorized action.");
+        }
     }
 
-    editAnswer(body: UpdateAnswerDto, id: number) {
-        const requestUrl = `answer/${id}`;
-        return axios.patch(requestUrl,body).then((response) => {return response.data;})
-        .catch(() => {
-            throw new  BadRequestException("Could not fetch data from the Data Layer.");
-        });
+    async editAnswer(headers:any, body: UpdateAnswerDto, ansid: number) {
+        let id : number = await verify(headers);
+
+        const requestUrl = `answer/${ansid}`;
+        const params = {
+            id : ansid,
+            user: true
+        };
+        let userId;
+        const answerGet = await axios.get('answer/one', { params })
+        .then((response) => { userId = response.data.user.id; })
+        .catch( () => {
+            throw new BadRequestException("Could not fetch data from the Data Layer.");
+        })
+
+        if(userId == id) {
+            return axios.patch(requestUrl,body).then((response) => {return response.data;})
+            .catch(() => {
+                throw new  BadRequestException("Could not fetch data from the Data Layer.");
+            });
+        }
+        else {
+            throw new UnauthorizedException("Unauthorized action.");
+        }
     }
 
-    deleteAnswer(id:number) {
-        const requestUrl = `answer/${id}`;
-        return axios.delete(requestUrl).then((response) => {return response.data;})
-        .catch(() => {
-            throw new  BadRequestException("Could not fetch data from the Data Layer.");
+    async deleteAnswer(headers:any, ansid:number) {
+        let id : number = await verify(headers);
+
+        const params = {
+            id : ansid,
+            user: true
+        };
+        let userId;
+        const answerGet = await axios.get('answer/one', { params })
+        .then((response) => { userId = response.data.user.id; })
+        .catch( () => {
+            throw new BadRequestException("Could not fetch data from the Data Layer.");
         });
+
+        if ( userId == id) {
+            const requestUrl = `answer/${ansid}`;
+            return axios.delete(requestUrl).then((response) => {return response.data;})
+            .catch(() => {
+                throw new  BadRequestException("Could not fetch data from the Data Layer.");
+            });
+        }
+        else {
+            throw new UnauthorizedException("Unauthorized action.");
+        }
+
     }
     
-    async vote(body: CreateUserAnswerVoteDto) {
+    async vote(headers:any, body: CreateUserAnswerVoteDto) {
+        let id : number = await verify(headers);
+
         const requestUrl = 'user-answer-vote';
         let userId, answerId,voteId,checkdata;
         let alrExists;
@@ -51,12 +98,15 @@ export class QuestionRunService {
             .catch(() => {
                 throw new  BadRequestException("Could not fetch data from the Data Layer.");
             });
-        }else {
+        }
+        else {
+            
             const reqAdder = `user-answer-vote/${voteId}`;
             const deleted = axios.delete(reqAdder).then((response) => { return response.data; })
             .catch(() => {
                 throw new BadRequestException("Could not fetch data from the Data Layer.");
             });
+            
             if(deleted) {
                 return axios.post(requestUrl,body).then((response) => {return response.data;})
                 .catch(() => {
@@ -69,11 +119,25 @@ export class QuestionRunService {
         }
     }
 
-    removeVote(id : number) {
-        const requestUrl = `user-answer-vote/${id}`;
-        return axios.delete(requestUrl).then((response) => { return response.data; })
+    async removeVote(headers:any, voteid : number) {
+        let id : number = await verify(headers);
+
+        const requestUrl = `user-answer-vote/${voteid}`;
+
+        let userId;
+        const thisVote = await axios.get(requestUrl).then((response) => { userId = response.data.userId; })
         .catch(() => {
             throw new  BadRequestException("Could not fetch data from the Data Layer.");
         });
+
+        if(id == userId) {
+            return axios.delete(requestUrl).then((response) => { return response.data; })
+            .catch(() => {
+                throw new  BadRequestException("Could not fetch data from the Data Layer.");
+            });
+        }
+        else {
+            throw new UnauthorizedException("Unauthorized action.");
+        }
     }
 }
